Refactored pro_gen
