+++ src/cpp/pro_gen/CMakeLists.txt
@@ -4,31 +4,23 @@ project(progen)
 
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
-find_package(Boost COMPONENTS program_options REQUIRED)
 
+find_package(Boost COMPONENTS program_options REQUIRED)
 find_package(nlohmann_json REQUIRED)
 
-file(GLOB pro_gen_headers CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/*.h)
-file(GLOB pro_gen_sources CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/*.cpp)
+file(GLOB_RECURSE pro_gen_headers CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/*.h)
+file(GLOB_RECURSE pro_gen_core CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/core/*.cpp)
+file(GLOB_RECURSE pro_gen_zombieworld CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/ZombieWorld/*.cpp)
+file(GLOB_RECURSE pro_gen_gridworld CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/GridWorld/*.cpp)
+file(GLOB_RECURSE pro_gen_dungeonworld CONFIGURE_DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/DungeonWorld/*.cpp)
 add_library(
   progen STATIC
   ${pro_gen_headers}
-  Pos.cpp
-  Object.cpp
-  AABB.cpp
-  World.cpp
-  Block.cpp
-  Door.cpp
-  Lever.cpp
-  Entity.cpp
-  Connection.cpp
-  ZombieworldPit.cpp
-  ZombieworldGroup.cpp
-  ZombieWorld.cpp
-  DungeonWorld.cpp
-  DungeonRoom1.cpp
-  DungeonRoom2.cpp
-  GridWorld.cpp)
+  ${pro_gen_core}
+  ${pro_gen_zombieworld}
+  ${pro_gen_gridworld}
+  ${pro_gen_dungeonworld}
+  )
 
 target_link_libraries(progen ${Boost_LIBRARIES} nlohmann_json::nlohmann_json)
 add_executable(generator generator.cpp)
