+++ src/cpp/pro_gen/DungeonWorld/aabbs/DungeonRoom1.h
@@ -1,12 +1,12 @@
 #pragma once
 
-#include "AABB.h"
+#include "../../core/AABB.h"
 
 /**
  * @brief This class represents a very specific building that is used in the
  *        Zombie mission
  */
-class DungeonRoom2 : public AABB {
+class DungeonRoom1 : public AABB {
   private:
     std::mt19937_64 gen;
 
@@ -25,7 +25,7 @@ class DungeonRoom2 : public AABB {
     void decorate();
 
     /**
-     * @brief Add some reward blocks (diamond) to the room
+     * @brief Add some reward blocks (gold) to the room
      */
     void addRewards();
 
@@ -42,16 +42,16 @@ class DungeonRoom2 : public AABB {
 
   public:
     /**
-     * @brief Construct a new Dungeon Room 2 object
+     * @brief Construct a new Dungeon Room 1 object
      *
      * @param id The id for this room (keep it unique for each instance)
      * @param topLeft The top left. Remember the y coordinate for this is the
      * lowest
      */
-    DungeonRoom2(std::string id, Pos& topLeft);
+    DungeonRoom1(std::string id, Pos& topLeft);
 
     /**
-     * @brief Destroy the Dungeon Room 2 object
+     * @brief Destroy the Zombieworld Group object.
      */
-    ~DungeonRoom2();
+    ~DungeonRoom1();
 };
