+++ src/cpp/pro_gen/core/Entity.cpp
